feat: added cpx, cpy, sep, jml, cmp, rep, jsr, jsl, ldy and ldx instructions, fixed internal register sizes to make sure no precision is lost, added zeropagey memory address mode

+++ src/cpu/bus.rs
@@ -12,7 +12,7 @@ impl Bus {
         }
     }
 
-    pub fn write_byte(&mut self, val: u8, addr: u16) {
+    pub fn write_byte(&mut self, val: u8, addr: u32) {
         match addr {
             _ => self.work_ram[addr as usize] = val,
         }
